third_party/WebKit/LayoutTests/virtual/mojo-loading/http/tests/security/media-element-audio-source-node-cross-origin-expected.txt

CONSOLE MESSAGE: MediaElementAudioSource outputs zeroes due to CORS access restrictions for http://localhost:8000/security/resources/webaudio/laughter.wav
CONSOLE ERROR: line 34: Uncaught (in promise) InvalidStateError: cannot startRendering when an OfflineAudioContext is running
MediaElementAudioSourceNode with cross-origin source

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS All samples correctly zeroed.
PASS successfullyParsed is true

TEST COMPLETE
